Remote all uses of Tullio.jl
The motivation for this is that package load time is super linear as a
function of number of packages/amount of code loaded.
For this reason, large packages with many dependencies can reach truly
extreme load times.
Not all packages display this problem, but Tullio
does. `@time_imports`, for example, shows:
```julia
  14424.8 ms  Tullio 0.30% compilation time
     56.5 ms  ParetoSmooth 53.39% compilation time (15% recompilation)
```
For a large package using `ParetoSmooth`. We would like to continue
using `ParetoSmooth`, but without taking this hit from `Tullio`.