Remove all uses of Tullio.jl (#72)
* Remote all uses of Tullio.jl

The motivation for this is that package load time is super linear as a
function of number of packages/amount of code loaded.
For this reason, large packages with many dependencies can reach truly
extreme load times.
Not all packages display this problem, but Tullio
does. `@time_imports`, for example, shows:
```julia
  14424.8 ms  Tullio 0.30% compilation time
     56.5 ms  ParetoSmooth 53.39% compilation time (15% recompilation)
```
For a large package using `ParetoSmooth`. We would like to continue
using `ParetoSmooth`, but without taking this hit from `Tullio`.

* Fix a test failure

* Fix test failure

* Bug fix

Co-authored-by: David Widmann <devmotion@users.noreply.github.com>

* Update src/ParetoSmooth.jl

Co-authored-by: David Widmann <devmotion@users.noreply.github.com>

* `@inbounds`, not `@inline` on loop

* Add `@inbounds` to enable SIMD

Additionally, allow hoisting constant polynomial coefficients out of loops.

* Bump version

So we can release.

* More `exp_inline`

---------

Co-authored-by: Carlos Parada <paradac@carleton.edu>
Co-authored-by: David Widmann <devmotion@users.noreply.github.com>